+++ app/CHANGELOG.md
@@ -16,7 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [DemoApp][Library] For `button` component, update to version 3.3.0 ([#832](https://github.com/Orange-OpenSource/ouds-flutter/issues/832))
 - [DemoApp][Library] Update `progress-indicator`: add size parameter to OudsCircularProgressIndicator for button integration ([#876](https://github.com/Orange-OpenSource/ouds-flutter/issues/876))
 ### Fixed
-- [Library] fix: fix Indicator Icon with text direction `list item` to resolve ar ([#925](https://github.com/Orange-OpenSource/ouds-flutter/issues/925))
+- [Library] `List item` with external indicator should have a role link ([#918](https://github.com/Orange-OpenSource/ouds-flutter/issues/918))
+- [Library] `List item` Trailing, labels and leading elements should be grouped ([#919](https://github.com/Orange-OpenSource/ouds-flutter/issues/919))
+- [Library] `List item` Leading/Trailing elements have a wrong accessible name ([#927](https://github.com/Orange-OpenSource/ouds-flutter/issues/927))
+- [Library] `List item` With 235% zoom, trailing text is truncated if displayed on 2 lines (iOS) ([#928](https://github.com/Orange-OpenSource/ouds-flutter/issues/928))
+- [Library] `list item` fix Indicator Icon with text direction  to resolve ar ([#925](https://github.com/Orange-OpenSource/ouds-flutter/issues/925))
 - [Library] Invalid value: Not in inclusive range 0..2: 3 dans OudsTabBar` ([#896](https://github.com/Orange-OpenSource/ouds-flutter/issues/896))
 - [DemoApp][Dependency] bump gradle-wrapper from 8.11.1 to 8.14.2 in /app/android ([#894](https://github.com/Orange-OpenSource/ouds-flutter/issues/894))